Revert "still need integration with mt"
This reverts commit d6c71f5bfa2e9afede0a039d88a19c6912c17b88.

wrong email adress

+++ src/dev/masters/services/IService.java
@@ -17,7 +17,7 @@ public interface IService<T> {
     public void ajouter(T t);
     public List<T> afficher()throws SQLException;
     public void supprimer(Long id) throws SQLException;
-    public void modifier(int id,T t) throws SQLException;
+    public void modifier(Long id,T t) throws SQLException;
 
 }
